defutil: Copy power pins placement status when applying DEF template
In newer OpenROAD, any power BTerm that isn't "FIXED" will be removed
completely if it's not connected when `pdngen` is called.

In OpenFrame those get connected later (in TinyTapeout case by some
custom script later and in the official example case using a macro).

So to work around that, we changed the DEF template to make sure all
power pins have the FIXED placement attribute and we make sure to copy
it when applying the template.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>