nix: Use custom OpenROAD version
We need a version recent enough to have the fix from OpenSTA to write
correct netlist when several power pins connect to the same net.

But old enough to not have the new incremental DRT that breaks everything.

We also need the new Antenna properties SWIG API to work with OpenLane2.

No official revision satisfy all of those, so I took a revision that
satisfy the first two and cherry-picked the third over it.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>